Make ColorColumn and CursorColumn visible
I'm not sure if this was intentional for minimalism reasons, so feel free to reject and I can maintain this as an override instead. But, I noticed the ColorColumn and CursorColumn ctermbg values were set to the same color as the background. Personally, I rely on the ColorColumn a lot.